requirements.txt
streamlit
openai
requests
python-dotenv
xhtml2pdf
